Web Inspector: Infrequent uncaught exception with debugger popovers breaks inspector https://bugs.webkit.org/show_bug.cgi?id=117755
Create the ZERO_SIZE and ZERO_RECT objects after we've setup the Size
and Rect prototypes, so they get the expected methods. We then need to
handle the ZERO_RECT case better, and not attempt to draw a background
in a canvas with a 0 size, which would produce an exception.

Patch by Joseph Pecoraro <pecoraro@apple.com> on 2013-06-18
Reviewed by Timothy Hatcher.

* UserInterface/Geometry.js:
* UserInterface/Popover.js:
(WebInspector.Popover.prototype._update):

Identifier: 135896@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151704 268f45cc-cd09-0410-ab3c-d52691b4dbfc