Sticky elements are missing or misplaced sometimes when reloading a page that scrolls to a named anchor https://bugs.webkit.org/show_bug.cgi?id=117819
Reviewed by Anders Carlsson.

At the end of FrameView::performPostLayoutTasks() we call scrollToAnchor() which attempts
to restore the scroll position to a named anchor, even when the document has been
changed by layout. This ends up in a call to ScrollView::scrollTo(), which in turn
calls repaintFixedElementsAfterScrolling(). However, repaintFixedElementsAfterScrolling()
would bail if m_nestedLayoutCount != 0, so we never updated layer positions which
depend on scroll position (but which would not be updated by layout, since their
renderers are not marked for layout when scrolling happens).

We've solved this problem once before in updateFixedElementsAfterScrolling() which
checks for m_nestedLayoutCount <= 1, so that we do work on the outermost nested layout.
Apply that same fix to repaintFixedElementsAfterScrolling().

Very timing-dependent, so hard to make a test.

* page/FrameView.cpp:
(WebCore::FrameView::repaintFixedElementsAfterScrolling):

Identifier: 135987@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151803 268f45cc-cd09-0410-ab3c-d52691b4dbfc