[BlackBerry] Send the correct offsets for misspelled words. https://bugs.webkit.org/show_bug.cgi?id=117846
JIRA116916.
When the caret is placed after a word, the offsets were calculated around
the proceeding space. As such, words that end a sentence worked correctly, but
midsentence words did not. The 'startOfWord' was calculated for the space ahead
and not for the word before it. Now finding the start of the word correctly and
calculating the end from there.

Patch by Nima Ghanavatian <nghanavatian@blackberry.com> on 2013-06-20
Reviewed by Rob Buis.

Internally reviewed by Genevieve Mak.

* WebKitSupport/InputHandler.cpp:
(BlackBerry::WebKit::InputHandler::requestSpellingCheckingOptions):

Identifier: 135992@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151809 268f45cc-cd09-0410-ab3c-d52691b4dbfc