Fix various crashes on sites with fixed backgrounds https://bugs.webkit.org/show_bug.cgi?id=117959
Source/WebCore:

Reviewed by Andy Estes.

FrameView::removeSlowRepaintObject() would assume that addSlowRepaintObject()
had been called before it, but this isn't always the case. For example, if
a page has a fixed background on the body, this falls into the accelerated
path in WK2 tiled mode, so addSlowRepaintObject() is never called. However,
we still call removeSlowRepaintObject() if the body is removed.

So null-check m_slowRepaintObjects to avoid crashing.

Test: platform/mac-wk2/tiled-drawing/fixed-background/fixed-background-removal.html

* page/FrameView.cpp:
(WebCore::FrameView::removeSlowRepaintObject):

LayoutTests:

Reviewed by Andy Estes.

Test that does a document.write on a page with a fixed background on the body.

* platform/mac-wk2/tiled-drawing/fixed-background/fixed-background-removal-expected.txt: Added.
* platform/mac-wk2/tiled-drawing/fixed-background/fixed-background-removal.html: Added.

Identifier: 136082@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151934 268f45cc-cd09-0410-ab3c-d52691b4dbfc