[GTK][WK2] SIGSEV in WebKit::WebPageContextMenuClient::customContextMenuItemSelected https://bugs.webkit.org/show_bug.cgi?id=116867
Reviewed by Anders Carlsson.

We did not reset the clients on WebPageProxy::close(), so right
clicking on a link after starting the navigation to some other
page was causing crashes because code was trying to access an
already freed m_contextMenuClient.

There were some other clients that could potentially trigger the
same issues. Reset them all to null to prevent this situation.

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::close):

Identifier: 136368@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@152367 268f45cc-cd09-0410-ab3c-d52691b4dbfc