[Autoconf] Define WTF_USE_OPENGL when OpenGL was found https://bugs.webkit.org/show_bug.cgi?id=118341
Reviewed by Gustavo Noronha Silva.

After r152275, WTF_USE_OPENGL is only defined if the accelerated compositing is enabled.
This breaks the build if disabling the accelerated compositing feature as the OpenGL-specific
code is not compiled anymore even if it's still required.

To avoid that, WTF_USE_OPENGL should be defined if OpenGL was found. The accelerated compositing
feature also relies on this condition, exiting the configuration process with an error otherwise.

* Source/autotools/SetupAutoconfHeader.m4:


Identifier: 136442@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@152454 268f45cc-cd09-0410-ab3c-d52691b4dbfc