[WK2][GTK] SIGSEV in webkitWebViewBaseSizeAllocate https://bugs.webkit.org/show_bug.cgi?id=118052
Patch by Alberto Garcia <berto@igalia.com> on 2013-07-15
Reviewed by Carlos Garcia Campos.

We don't need to check whether the drawing area is empty or not in
order to set needsResizeOnMap.

It can also happen that the drawing area pointer is null (if
e.g. the web process crashes), which is the reason for this
SIGSEGV.

A test to prevent this was added in r88646 but was later moved to
resizeWebKitWebViewBaseFromAllocation().

* UIProcess/API/gtk/WebKitWebViewBase.cpp:
(webkitWebViewBaseSizeAllocate):

Identifier: 136562@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@152626 268f45cc-cd09-0410-ab3c-d52691b4dbfc