Make atob() / btoa() argument non optional https://bugs.webkit.org/show_bug.cgi?id=118844
Reviewed by Kentaro Hara.

Source/WebCore:

According to the latest specification, the argument to atob() / btoa()
should not be optional:
http://www.whatwg.org/specs/web-apps/current-work/multipage/webappapis.html#windowbase64

This patch makes WebKit behave according to the specification. The argument
is also mandatory in Firefox, IE10 and Blink.

atob() / btoa() are also moved to their own WindowBase64 interface which
the Window interface implements. This does not change the behavior but
this is closer to the specification and it will make exposing those
methods to workers easier later on.

No new tests, already covered by:
fast/dom/Window/atob-btoa.html

* CMakeLists.txt:
* DerivedSources.make:
* DerivedSources.pri:
* GNUmakefile.list.am:
* page/DOMWindow.idl:
* page/WindowBase64.idl: Added.

LayoutTests:

Update and rebaseline fast/dom/Window/atob-btoa.html as atob() /
btoa() 's behavior has changed to match the specification and
the behavior in other browsers.

* fast/dom/Window/atob-btoa-expected.txt:
* fast/dom/Window/atob-btoa.html:


Identifier: 136720@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@152859 268f45cc-cd09-0410-ab3c-d52691b4dbfc