fourthTier: Segfault in jsc with simple test program when running with profile dumping enabled https://bugs.webkit.org/show_bug.cgi?id=116082
It's crashing because CodeBlock::baselineVersion() doesn't know how to handle the case where 'this' is the
baseline version but it hasn't been assigned to the m_blahCodeBlock field in BlahExecutable. The fix is to
check if we're the baseline version in baselineVersion() and return this if so.

Reviewed by Filip Pizlo.

* bytecode/CodeBlock.h:
(JSC::CodeBlock::baselineVersion):

Identifier: 136943@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153159 268f45cc-cd09-0410-ab3c-d52691b4dbfc